requirements.txt
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
APScheduler==3.10.4
argcomplete==3.6.2
async-timeout==5.0.1
attrs==25.3.0
azure==4.0.0
azure-applicationinsights==0.1.1
azure-batch==4.1.3
azure-common==1.1.28
azure-core==1.35.1
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-datalake-store==0.0.53
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.1
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.8
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
breadability==0.1.20
Brotli==1.1.0
certifi==2025.7.14
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorhash==2.1.0
configparser==7.2.0
contourpy==1.3.0
cryptography==45.0.7
cssselect2==0.8.0
cycler==0.12.1
decouple==0.0.7
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docx==0.2.4
dominate==2.9.1
dotenv==0.9.9
email-validator==2.2.0
escpos==1.9
et-xmlfile==2.0.0
exceptiongroup==1.3.0
faker==37.5.3
flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
flask-monitoringdashboard==4.0.4
Flask-Principal==0.4.0
flask-security==5.6.2
Flask-SocketIO==5.5.1
flask-sqlalchemy==3.1.1
Flask-Uploads==0.2.1
Flask-User==1.0.2.2
flask-wtf==1.2.2
fonttools==4.59.2
fpdf==1.7.2
frozenlist==1.7.0
future==1.0.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.7
lxml==6.0.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
msal==1.34.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.6.3
narwhals==2.1.2
nltk==3.9.1
numpy==2.0.2
oauthlib==3.3.1
ollama==0.5.3
openai==1.107.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
psutil==7.1.0
psycopg2-binary==2.9.11
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.7
pydantic-core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pyphen==0.17.2
pyserial==3.5
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.2
qrcode==8.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.0
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.13.1
simple-websocket==1.1.0
six==1.17.0
sklearn==0.0
sniffio==1.3.1
sqlalchemy==2.0.40
sumy==0.11.0
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
twilio==9.7.0
typing-extensions==4.13.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==2.0.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
viivakoodi==0.8.0
visitor==0.1.3
weasyprint==66.0
webencodings==0.5.1
werkzeug==3.1.3
wsproto==1.2.0
wtforms==3.2.1
XlsxWriter==3.2.3
yarl==1.20.1
zipp==3.21.0
zopfli==0.2.3.post1
